Safari-609.1.11.1/LayoutTests/accessibility/table-cell-display-block-expected.txt

This tests that if a table cell uses display:block, the table cell will still be accessible

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cell1a.isEqual(cell1b) is true
PASS cell1a.role is 'AXRole: AXCell'
PASS domCell1.style.display is 'block'
PASS table2.role is 'AXRole: AXGroup'
PASS cell2.role == 'AXRole: AXCell' is false
PASS successfullyParsed is true

TEST COMPLETE
